refactor: lift the menu, window and auto-updater out of the entry point
