+++ src/main/index.js
@@ -3,7 +3,7 @@ if (process.env.NODE_ENV !== "production") {
   require("dotenv").config()
 }
 
-const { app, BrowserWindow, Menu, shell, dialog } = require("electron")
+const { app, BrowserWindow, shell, dialog } = require("electron")
 const { autoUpdater } = require("electron-updater")
 const path = require("path")
 const isDev = process.env.NODE_ENV === "development"
@@ -16,20 +16,28 @@ const { PotInstaller } = require("./services/pot-installer")
 const { SettingsStore } = require("./services/settings-store")
 const { Analytics } = require("./services/analytics")
 const IPCHandlers = require("./ipc-handlers")
-const { APP_CONFIG } = require("./utils/constants")
 const {
   describeError,
   getAppVersion,
   isFirstLaunch
 } = require("./utils/analytics-helpers")
+const { createMenu } = require("./menu")
+const {
+  setupAutoUpdater,
+  checkForUpdatesWithRetry,
+  downloadUpdateWithRetry,
+  setupPeriodicUpdateChecks
+} = require("./auto-updater")
+const {
+  giveYouTubeEmbedsAReferer,
+  createWindow,
+  loadApplication,
+  onWindowClosed,
+  getAppIcon
+} = require("./window")
 
 // let the user get their first download going before we check for updates
 const UPDATE_CHECK_DELAY_MS = 90 * 1000
-
-// the site the embed player is told it is embedded on. it has to name a real
-// site that is not youtube itself - a youtube.com referer is refused as a
-// self-referential embed (error 152). see giveYouTubeEmbedsAReferer()
-const EMBED_REFERER = "https://cliply.space/"
 
 // what getAppVersion() reports when it cannot read package.json. persisting it
 // would poison the launch after this one: it comes back as previous_version,
@@ -47,169 +55,6 @@ const QUIT_FLUSH_TIMEOUT_MS = 2000
  * exists to explain - including "busy", where the check never ran at all.
  */
 const ENGINE_CURRENT_REASONS = new Set(["up-to-date", "completed"])
-
-/**
- * what the native menu says, in the two languages the app speaks
- *
- * this one surface follows app.getLocale() - the OS - rather than the in-app
- * language toggle. the menu bar is the system's, it is built before a renderer
- * exists to ask, and the alternative is an ipc round trip on every startup.
- *
- * sentence case here, unlike the lowercase in-app voice: these labels sit in
- * the OS's own menu bar next to the OS's own words. product names stay latin.
- */
-const MENU_TEXT = {
-  en: {
-    file: "File",
-    newDownload: "New Download",
-    openDownloads: "Open Downloads Folder",
-    quit: "Quit",
-
-    edit: "Edit",
-    undo: "Undo",
-    redo: "Redo",
-    cut: "Cut",
-    copy: "Copy",
-    paste: "Paste",
-
-    tools: "Tools",
-    checkUpdates: "Check for Updates",
-    videoEngine: "Video engine",
-    versionUnknown: "version unknown",
-    sendUsageData: "Send usage data",
-
-    view: "View",
-    reload: "Reload",
-    forceReload: "Force Reload",
-    devTools: "Toggle Developer Tools",
-    actualSize: "Actual Size",
-    zoomIn: "Zoom In",
-    zoomOut: "Zoom Out",
-    fullscreen: "Toggle Fullscreen",
-
-    window: "Window",
-    minimize: "Minimize",
-    close: "Close",
-    zoom: "Zoom",
-    bringAllToFront: "Bring All to Front",
-
-    help: "Help",
-    aboutCliply: "About Cliply",
-    systemHealth: "System Health",
-    reportIssue: "Report Issue",
-
-    // the macos app menu builds two of its labels around the app's own name
-    about: "About",
-    services: "Services",
-    hide: "Hide",
-    hideOthers: "Hide Others",
-    showAll: "Show All",
-
-    version: "Version",
-    aboutBlurb:
-      "Your fave little desktop app; powered by open source tools (>ᴗ•)",
-
-    // the boxes these menu items put up when something goes wrong
-    error: "Error",
-    updateCheckFailed: "Update Check Failed",
-    updateCheckFailedMessage: "Failed to check for updates.",
-    updateCheckFailedDetail: "Please try again later.",
-    updateUnavailable: "Update Check Unavailable",
-    updateUnavailableMessage: "Update checking is not available.",
-    updateUnavailableDetail:
-      "Updates are only available in production builds.",
-    prefNotSaved: "Couldn't save that preference",
-    prefNotSavedMessage: "Your analytics preference could not be saved.",
-    prefNotSavedDetail: "Please try again.",
-    healthCheckFailed: "Failed to check system health.",
-
-    healthHealthy: "System Status: Healthy",
-    healthError: "System Status: Error",
-    healthDownloader: "Downloader",
-    healthUnknown: "Unknown",
-    healthFound: "Found",
-    healthMissing: "Missing",
-    healthValid: "Valid",
-    healthInvalid: "Invalid",
-    healthActive: "Active downloads",
-    healthUptime: "Uptime",
-    healthMinutes: "minutes"
-  },
-  ru: {
-    file: "Файл",
-    newDownload: "Новая загрузка",
-    openDownloads: "Открыть папку загрузок",
-    quit: "Выход",
-
-    edit: "Правка",
-    undo: "Отменить",
-    redo: "Повторить",
-    cut: "Вырезать",
-    copy: "Копировать",
-    paste: "Вставить",
-
-    tools: "Инструменты",
-    checkUpdates: "Проверить обновления",
-    videoEngine: "Движок видео",
-    versionUnknown: "версия неизвестна",
-    sendUsageData: "Отправлять данные об использовании",
-
-    view: "Вид",
-    reload: "Перезагрузить",
-    forceReload: "Перезагрузить принудительно",
-    devTools: "Инструменты разработчика",
-    actualSize: "Исходный размер",
-    zoomIn: "Увеличить",
-    zoomOut: "Уменьшить",
-    fullscreen: "Полноэкранный режим",
-
-    window: "Окно",
-    minimize: "Свернуть",
-    close: "Закрыть",
-    zoom: "Масштаб",
-    bringAllToFront: "Все окна на передний план",
-
-    help: "Справка",
-    aboutCliply: "О Cliply",
-    systemHealth: "Состояние системы",
-    reportIssue: "Сообщить о проблеме",
-
-    about: "О",
-    services: "Службы",
-    hide: "Скрыть",
-    hideOthers: "Скрыть остальные",
-    showAll: "Показать все",
-
-    version: "Версия",
-    aboutBlurb:
-      "Ваше любимое маленькое приложение; работает на инструментах с открытым кодом (>ᴗ•)",
-
-    error: "Ошибка",
-    updateCheckFailed: "Не удалось проверить обновления",
-    updateCheckFailedMessage: "Не удалось проверить наличие обновлений.",
-    updateCheckFailedDetail: "Попробуйте позже.",
-    updateUnavailable: "Проверка обновлений недоступна",
-    updateUnavailableMessage: "Проверка обновлений здесь не работает.",
-    updateUnavailableDetail:
-      "Обновления доступны только в установленной версии.",
-    prefNotSaved: "Не удалось сохранить настройку",
-    prefNotSavedMessage: "Настройку аналитики не удалось сохранить.",
-    prefNotSavedDetail: "Попробуйте ещё раз.",
-    healthCheckFailed: "Не удалось проверить состояние системы.",
-
-    healthHealthy: "Состояние системы: в порядке",
-    healthError: "Состояние системы: ошибка",
-    healthDownloader: "Загрузчик",
-    healthUnknown: "неизвестно",
-    healthFound: "найден",
-    healthMissing: "отсутствует",
-    healthValid: "рабочие",
-    healthInvalid: "нерабочие",
-    healthActive: "Активные загрузки",
-    healthUptime: "Время работы",
-    healthMinutes: "мин."
-  }
-}
 
 class CliplyApp {
   constructor() {
@@ -486,199 +331,20 @@ class CliplyApp {
     })
   }
 
-  // setup auto-updater
   setupAutoUpdater() {
-    try {
-      // configure auto-updater
-      autoUpdater.checkForUpdatesAndNotify = false
-      autoUpdater.autoDownload = false
-      autoUpdater.autoInstallOnAppQuit = false
-
-      // disable code signature verification for unsigned builds
-      autoUpdater.verifyUpdateCodeSignature = false
-
-      // checking for updates
-      autoUpdater.on("checking-for-update", () => {
-        if (this.mainWindow && !this.mainWindow.isDestroyed()) {
-          this.mainWindow.webContents.send("update:checking")
-        }
-      })
-
-      // update available - handle based on platform
-      autoUpdater.on("update-available", (info) => {
-        const isMac = process.platform === "darwin"
-
-        if (isMac) {
-          console.log(
-            "Update available:",
-            info.version,
-            "- showing manual download for macOS"
-          )
-
-          // macOS: Show manual download popup
-          if (this.mainWindow && !this.mainWindow.isDestroyed()) {
-            this.mainWindow.webContents.send("update:available", {
-              version: info.version,
-              releaseNotes: info.releaseNotes,
-              releaseDate: info.releaseDate,
-              requiresManualDownload: true,
-              platform: "darwin"
-            })
-          }
-        } else {
-          console.log(
-            "Update available:",
-            info.version,
-            "- auto-downloading..."
-          )
-
-          // Windows/Linux: Auto-download as before
-          if (this.mainWindow && !this.mainWindow.isDestroyed()) {
-            this.mainWindow.webContents.send("update:available", {
-              version: info.version,
-              releaseNotes: info.releaseNotes,
-              releaseDate: info.releaseDate,
-              autoDownloading: true
-            })
-          }
-
-          // auto-download for non-macOS platforms
-          this.downloadUpdateWithRetry().catch((error) => {
-            console.error("Auto-download failed:", error)
-          })
-        }
-      })
-
-      // update not available
-      autoUpdater.on("update-not-available", () => {
-        if (this.mainWindow && !this.mainWindow.isDestroyed()) {
-          this.mainWindow.webContents.send("update:not-available")
-        }
-      })
-
-      // download progress
-      autoUpdater.on("download-progress", (progress) => {
-        if (this.mainWindow && !this.mainWindow.isDestroyed()) {
-          this.mainWindow.webContents.send("update:download-progress", {
-            percent: Math.round(progress.percent),
-            bytesPerSecond: progress.bytesPerSecond,
-            total: progress.total,
-            transferred: progress.transferred
-          })
-        }
-      })
-
-      // update downloaded - enable auto-install on quit for all updates
-      autoUpdater.on("update-downloaded", (info) => {
-        console.log("Update downloaded:", info.version, "- ready to install")
-
-        // enable auto-install on quit for all updates
-        autoUpdater.autoInstallOnAppQuit = true
-
-        // notify renderer that update is ready
-        if (this.mainWindow && !this.mainWindow.isDestroyed()) {
-          this.mainWindow.webContents.send("update:downloaded", {
-            version: info.version,
-            autoInstallOnQuit: true
-          })
-        }
-      })
-
-      // error
-      autoUpdater.on("error", (error) => {
-        console.error("Auto-updater error:", error.message)
-        if (this.mainWindow && !this.mainWindow.isDestroyed()) {
-          this.mainWindow.webContents.send("update:error", {
-            message: error.message
-          })
-        }
-      })
-
-      // check for updates after app ready
-      app.whenReady().then(() => {
-        const shouldCheck = isDev || Math.random() < 0.9
-
-        if (shouldCheck) {
-          setTimeout(() => {
-            this.checkForUpdatesWithRetry().catch((error) => {
-              console.error("Failed to check for updates:", error)
-            })
-          }, 3000)
-        }
-
-        // setup periodic update checks every 12 hours
-        this.setupPeriodicUpdateChecks()
-      })
-    } catch (error) {
-      console.error("Auto-updater setup failed:", error)
-    }
+    return setupAutoUpdater(this)
   }
 
-  // retry logic for update checks
-  async checkForUpdatesWithRetry(
-    maxRetries = APP_CONFIG.UPDATE_CONFIG.MAX_CHECK_RETRIES
-  ) {
-    for (let attempt = 1; attempt <= maxRetries; attempt++) {
-      try {
-        await autoUpdater.checkForUpdates()
-        return
-      } catch (error) {
-        console.error(`Update check attempt ${attempt} failed:`, error.message)
-
-        if (attempt === maxRetries) {
-          throw error
-        }
-
-        // wait before retry (exponential backoff)
-        const delay = Math.min(1000 * Math.pow(2, attempt - 1), 10000)
-        await new Promise((resolve) => setTimeout(resolve, delay))
-      }
-    }
+  checkForUpdatesWithRetry(maxRetries) {
+    return checkForUpdatesWithRetry(maxRetries)
   }
 
-  // retry logic for update downloads
-  async downloadUpdateWithRetry(
-    maxRetries = APP_CONFIG.UPDATE_CONFIG.MAX_DOWNLOAD_RETRIES
-  ) {
-    for (let attempt = 1; attempt <= maxRetries; attempt++) {
-      try {
-        await autoUpdater.downloadUpdate()
-        return
-      } catch (error) {
-        console.error(
-          `Update download attempt ${attempt} failed:`,
-          error.message
-        )
-
-        if (attempt === maxRetries) {
-          throw error
-        }
-
-        // wait before retry
-        const delay = 2000 * attempt
-        await new Promise((resolve) => setTimeout(resolve, delay))
-      }
-    }
+  downloadUpdateWithRetry(maxRetries) {
+    return downloadUpdateWithRetry(maxRetries)
   }
 
-  // setup periodic update checks
   setupPeriodicUpdateChecks() {
-    // check every 12 hours using config
-    const checkInterval = APP_CONFIG.UPDATE_CONFIG.PERIODIC_CHECK_INTERVAL
-
-    // unref'd like every other timer here: a 12-hour interval is the longest
-    // lived handle in the process, and an armed one is on its own enough to
-    // hold the event loop open long after there is anything left to do
-    this.periodicUpdateTimer = setInterval(() => {
-      // only check if app is not quitting and in production
-      if (!this.isQuitting && !isDev) {
-        console.log("Performing periodic update check...")
-        this.checkForUpdatesWithRetry().catch((error) => {
-          console.error("Periodic update check failed:", error)
-        })
-      }
-    }, checkInterval)
-    this.periodicUpdateTimer.unref()
+    return setupPeriodicUpdateChecks(this)
   }
 
   // setup app event handlers
@@ -806,150 +472,20 @@ class CliplyApp {
       })
   }
 
-  /**
-   * give youtube's embed player a referer to read
-   *
-   * the packaged app is served with loadFile(), so the renderer runs under
-   * file:// - and chromium sends no Referer at all from a file:// page. the
-   * embed player decides playability from that header server-side: without it
-   * the response carries an errorScreen instead of a video, which is what the
-   * user sees as error 153. dev never hits it because vite serves the same
-   * page over http://localhost.
-   *
-   * scoped to the embed url so no other request the app makes gains a referer
-   * it did not have before.
-   */
   giveYouTubeEmbedsAReferer() {
-    this.mainWindow.webContents.session.webRequest.onBeforeSendHeaders(
-      { urls: ["https://www.youtube.com/embed/*"] },
-      (details, callback) => {
-        callback({
-          requestHeaders: {
-            ...details.requestHeaders,
-            Referer: EMBED_REFERER
-          }
-        })
-      }
-    )
+    return giveYouTubeEmbedsAReferer(this)
   }
 
-  // create main window
   createWindow() {
-    // create browser window
-    this.mainWindow = new BrowserWindow({
-      width: 1200,
-      height: 800,
-      minWidth: 800,
-      minHeight: 600,
-      show: false,
-      webPreferences: {
-        nodeIntegration: false,
-        contextIsolation: true,
-        enableRemoteModule: false,
-        webSecurity: false, // for youtube iframe compatibility
-        allowRunningInsecureContent: true,
-        preload: path.join(__dirname, "..", "preload", "preload.js"),
-        sandbox: false,
-        experimentalFeatures: false,
-        enableBlinkFeatures: "",
-        disableBlinkFeatures: "Auxclick"
-      },
-      icon: this.getAppIcon(),
-      titleBarStyle: process.platform === "darwin" ? "hiddenInset" : "default"
-    })
-
-    this.giveYouTubeEmbedsAReferer()
-
-    // window event handlers
-    this.mainWindow.on("closed", this.onWindowClosed)
-
-    this.mainWindow.on("close", (_event) => {
-      // allow close during update
-      if (global.isUpdating) {
-        return
-      }
-
-      // quit the app when close button is clicked (consistent behavior)
-      if (!this.isQuitting) {
-        this.isQuitting = true
-        app.quit()
-      }
-    })
-
-    // show window when ready
-    this.mainWindow.once("ready-to-show", () => {
-      this.mainWindow.show()
-
-      if (isDev) {
-        this.mainWindow.webContents.openDevTools()
-      }
-    })
-
-    // load the app
-    this.loadApplication()
-
-    // set ipc handlers main window reference
-    if (this.ipcHandlers) {
-      this.ipcHandlers.setMainWindow(this.mainWindow)
-    }
+    return createWindow(this)
   }
 
-  // load application ui
   loadApplication() {
-    if (isDev) {
-      // development: load from vite dev server
-      const startUrl = "http://localhost:5173"
-      this.mainWindow.loadURL(startUrl)
-    } else {
-      // production: load from packaged files
-      const rendererPath = path.join(
-        __dirname,
-        "renderer",
-        "dist",
-        "index.html"
-      )
-
-      try {
-        this.mainWindow.loadFile(rendererPath).catch((error) => {
-          console.error(`Failed to load renderer: ${error.message}`)
-
-          // fallback path
-          const fallbackPath = path.join(
-            process.resourcesPath,
-            "app.asar",
-            "src",
-            "main",
-            "renderer",
-            "dist",
-            "index.html"
-          )
-
-          return this.mainWindow
-            .loadFile(fallbackPath)
-            .catch((fallbackError) => {
-              console.error(`Fallback also failed: ${fallbackError.message}`)
-
-              dialog.showErrorBox(
-                "Application Error",
-                `Failed to load the application interface.\n\nPrimary path: ${rendererPath}\nFallback path: ${fallbackPath}\n\nPlease reinstall the application.`
-              )
-              app.quit()
-            })
-        })
-      } catch (error) {
-        console.error(`Critical error loading application: ${error.message}`)
-        dialog.showErrorBox(
-          "Critical Error",
-          `Cannot start application: ${error.message}`
-        )
-        app.quit()
-      }
-    }
+    return loadApplication(this)
   }
 
-  // handle main window closed
   onWindowClosed() {
-    this.mainWindow = null
+    return onWindowClosed(this)
   }
 
   /**
@@ -1062,328 +598,12 @@ class CliplyApp {
     app.quit()
   }
 
-  // get app icon path
   getAppIcon() {
-    const iconName = process.platform === "win32" ? "icon.ico" : "icon.png"
-    return path.join(__dirname, "..", "..", "assets", iconName)
+    return getAppIcon()
   }
 
-  // create application menu
   createMenu() {
-    // synchronous on purpose: the menu is built during startup and again after
-    // an update lands, and neither moment can wait on a --version probe
-    const engineVersion = this.services.ytdlpEngine.getKnownVersion()
-
-    // the OS's language, not the in-app toggle. both callers run after ready,
-    // where getLocale() has settled - and a locale we cannot read leaves an
-    // english menu, which is better than an initialisation that aborts here
-    let locale = ""
-    try {
-      locale = app.getLocale()
-    } catch (error) {
-      console.error("Could not read the OS locale:", error)
-    }
-
-    const T =
-      MENU_TEXT[
-        typeof locale === "string" && locale.toLowerCase().startsWith("ru")
-          ? "ru"
-          : "en"
-      ]
-
-    const template = [
-      {
-        label: T.file,
-        submenu: [
-          {
-            label: T.newDownload,
-            accelerator: "CmdOrCtrl+N",
-            click: () => {
-              if (this.mainWindow) {
-                this.mainWindow.webContents.send("menu:new-download")
-              }
-            }
-          },
-          { type: "separator" },
-          {
-            label: T.openDownloads,
-            accelerator: "CmdOrCtrl+D",
-            click: async () => {
-              try {
-                if (this.ipcHandlers) {
-                  await this.ipcHandlers.handleOpenDownloadFolder()
-                }
-              } catch (error) {
-                console.error("Failed to open downloads folder:", error)
-              }
-            }
-          },
-          { type: "separator" },
-          {
-            label: T.quit,
-            accelerator: process.platform === "darwin" ? "Cmd+Q" : "Ctrl+Q",
-            click: () => {
-              app.quit()
-            }
-          }
-        ]
-      },
-      {
-        label: T.edit,
-        submenu: [
-          { label: T.undo, accelerator: "CmdOrCtrl+Z", role: "undo" },
-          { label: T.redo, accelerator: "Shift+CmdOrCtrl+Z", role: "redo" },
-          { type: "separator" },
-          { label: T.cut, accelerator: "CmdOrCtrl+X", role: "cut" },
-          { label: T.copy, accelerator: "CmdOrCtrl+C", role: "copy" },
-          { label: T.paste, accelerator: "CmdOrCtrl+V", role: "paste" }
-        ]
-      },
-      {
-        label: T.tools,
-        submenu: [
-          {
-            label: T.checkUpdates,
-            click: async () => {
-              try {
-                if (this.ipcHandlers) {
-                  const result = await this.ipcHandlers.handleCheckForUpdates()
-
-                  if (result.success) {
-                    // update notification component handles ui feedback
-                  } else {
-                    console.error("Update check failed:", result.error?.message)
-                    dialog.showMessageBox(this.mainWindow, {
-                      type: "error",
-                      title: T.updateCheckFailed,
-                      message: T.updateCheckFailedMessage,
-                      detail:
-                        result.error?.message || T.updateCheckFailedDetail,
-                      buttons: ["OK"]
-                    })
-                  }
-                } else {
-                  dialog.showMessageBox(this.mainWindow, {
-                    type: "warning",
-                    title: T.updateUnavailable,
-                    message: T.updateUnavailableMessage,
-                    detail: T.updateUnavailableDetail,
-                    buttons: ["OK"]
-                  })
-                }
-              } catch (error) {
-                console.error("Manual update check failed:", error)
-                dialog.showErrorBox(T.error, T.updateCheckFailedMessage)
-              }
-            }
-          },
-          { type: "separator" },
-          {
-            // which engine this install actually downloads with - the single
-            // most useful thing to know before filing an issue, and the same
-            // string the report attaches.
-            //
-            // read from the engine rather than from analytics: telemetry is
-            // something the user can switch off, and this line has to stay
-            // right when they do. it is deliberately unknown-tolerant - a
-            // failed probe leaves us with no version for the whole run, and
-            // guessing one would be worse than admitting it
-            label: engineVersion
-              ? `${T.videoEngine}: yt-dlp ${engineVersion}`
-              : `${T.videoEngine}: ${T.versionUnknown}`,
-            enabled: false
-          },
-          { type: "separator" },
-          {
-            // deliberately not "anonymous". every event carries a persistent
-            // install id and a city derived from the ip, which is pseudonymous
-            // - and this label is read far more often than PRIVACY.md, so it
-            // is where the word would do its damage
-            label: T.sendUsageData,
-            type: "checkbox",
-            checked: this.services.analytics.isEnabled(),
-            click: async (menuItem) => {
-              // through the service, never the store: the opt-out gate is only
-              // re-read at init(), so writing the preference behind its back
-              // leaves this session sending for the rest of its life
-              const result = await this.services.analytics.setEnabled(
-                menuItem.checked
-              )
-
-              // this handler closed over the live MenuItem, and electron neither
-              // serialises clicks nor disables the item while one is running -
-              // so an older click resuming late acts on the checkbox the NEWEST
-              // click left behind, not on its own. a result the service has
-              // marked superseded has nothing to say about it, and a dialog
-              // naming a click already replaced is noise with no way to tell
-              // which click it was about.
-              if (result && result.superseded) {
-                return
-              }
-
-              // a privacy control that silently fails to persist would come
-              // back on at the next launch - say so rather than pretend
-              if (result && result.success === false) {
-                // read from the service, never inverted from the click. a tick
-                // derived from isEnabled() cannot disagree with what is
-                // actually being sent; one that flips agrees only as long as
-                // every path gets the arithmetic right, and it has been wrong
-                // twice. note a failed opt-out therefore leaves the tick OFF -
-                // the service is inert for this session either way, and it is
-                // the dialog that says it will not survive a restart.
-                menuItem.checked = this.services.analytics.isEnabled()
-                dialog.showMessageBox(this.mainWindow, {
-                  type: "error",
-                  title: T.prefNotSaved,
-                  message: T.prefNotSavedMessage,
-                  detail: result.error || T.prefNotSavedDetail,
-                  buttons: ["OK"]
-                })
-              }
-            }
-          }
-        ]
-      },
-      {
-        label: T.view,
-        submenu: [
-          { label: T.reload, accelerator: "CmdOrCtrl+R", role: "reload" },
-          {
-            label: T.forceReload,
-            accelerator: "CmdOrCtrl+Shift+R",
-            role: "forceReload"
-          },
-          {
-            label: T.devTools,
-            accelerator: "F12",
-            role: "toggleDevTools"
-          },
-          { type: "separator" },
-          {
-            label: T.actualSize,
-            accelerator: "CmdOrCtrl+0",
-            role: "resetZoom"
-          },
-          { label: T.zoomIn, accelerator: "CmdOrCtrl+Plus", role: "zoomIn" },
-          { label: T.zoomOut, accelerator: "CmdOrCtrl+-", role: "zoomOut" },
-          { type: "separator" },
-          {
-            label: T.fullscreen,
-            accelerator: "F11",
-            role: "togglefullscreen"
-          }
-        ]
-      },
-      {
-        label: T.window,
-        submenu: [
-          { label: T.minimize, accelerator: "CmdOrCtrl+M", role: "minimize" },
-          { label: T.close, accelerator: "CmdOrCtrl+W", role: "close" }
-        ]
-      },
-      {
-        label: T.help,
-        submenu: [
-          {
-            label: T.aboutCliply,
-            click: () => {
-              dialog.showMessageBox(this.mainWindow, {
-                type: "info",
-                title: T.aboutCliply,
-                message: "Cliply Desktop",
-                detail: `${T.version}: ${getAppVersion()}\n\n${T.aboutBlurb}`,
-                buttons: ["OK"]
-              })
-            }
-          },
-          {
-            label: T.systemHealth,
-            click: async () => {
-              try {
-                if (this.ipcHandlers) {
-                  const health = await this.ipcHandlers.handleSystemHealth()
-
-                  const message = health.success
-                    ? `${T.healthHealthy}\n\n${T.healthDownloader}: ${
-                        health.data.engine.version || T.healthUnknown
-                      }\nFFmpeg: ${
-                        health.data.engine.ffmpeg
-                          ? T.healthFound
-                          : T.healthMissing
-                      }\nCookies: ${
-                        health.data.cookies.hasValid
-                          ? T.healthValid
-                          : T.healthInvalid
-                      }\n${T.healthActive}: ${
-                        health.data.downloads.active
-                      }\n${T.healthUptime}: ${Math.floor(
-                        health.data.performance.uptime / 60
-                      )} ${T.healthMinutes}`
-                    : `${T.healthError}\n\n${health.error.message}`
-
-                  dialog.showMessageBox(this.mainWindow, {
-                    type: health.success ? "info" : "error",
-                    title: T.systemHealth,
-                    message,
-                    buttons: ["OK"]
-                  })
-                }
-              } catch (error) {
-                console.error("System health check failed:", error)
-                dialog.showErrorBox(T.error, T.healthCheckFailed)
-              }
-            }
-          },
-          { type: "separator" },
-          {
-            label: T.reportIssue,
-            click: () => {
-              shell.openExternal("https://github.com/Cliply/Cliply/issues")
-            }
-          }
-        ]
-      }
-    ]
-
-    // macos menu adjustments
-    if (process.platform === "darwin") {
-      template.unshift({
-        label: app.getName(),
-        submenu: [
-          { label: T.about + " " + app.getName(), role: "about" },
-          { type: "separator" },
-          { label: T.services, role: "services", submenu: [] },
-          { type: "separator" },
-          {
-            label: T.hide + " " + app.getName(),
-            accelerator: "Command+H",
-            role: "hide"
-          },
-          {
-            label: T.hideOthers,
-            accelerator: "Command+Shift+H",
-            role: "hideothers"
-          },
-          { label: T.showAll, role: "unhide" },
-          { type: "separator" },
-          { label: T.quit, accelerator: "Command+Q", click: () => app.quit() }
-        ]
-      })
-
-      // window menu
-      template[5].submenu = [
-        { label: T.close, accelerator: "CmdOrCtrl+W", role: "close" },
-        { label: T.minimize, accelerator: "CmdOrCtrl+M", role: "minimize" },
-        { label: T.zoom, role: "zoom" },
-        { type: "separator" },
-        { label: T.bringAllToFront, role: "front" }
-      ]
-    }
-
-    // kept so noteEngineVersion can tell "the menu is up and now reads wrong"
-    // from "startup has not built one yet"
-    this.menu = Menu.buildFromTemplate(template)
-    Menu.setApplicationMenu(this.menu)
+    return createMenu(this)
   }
 }
